Change Redux directory structure
