--- src/slices/index.tsx
@@ -1,10 +1,0 @@
-import { combineReducers } from "@reduxjs/toolkit";
-
-import sceneReducer from "./scene";
-
-const rootReducer = combineReducers({
-  sceneReducer,
-})
-export type RootState = ReturnType<typeof rootReducer>
-
-export default rootReducer
